+++ packages/tests/src/e2e/vs/BlazorAppHappyPath.tests.ts
