+++ packages/tests/src/e2e/vs/MessageExtensionActionHappyPath.dotnet.tests.ts
@@ -4,9 +4,9 @@
 /**
  * @author Siglud <fanhu@microsoft.com>
  **/
-import { happyPathTest } from "./BotHappyPathCommon";
-import { Runtime } from "../../commonlib/constants";
 import { it } from "@microsoft/extra-shot-mocha";
+import { Runtime } from "../../commonlib/constants";
+import { happyPathTest } from "../bot/BotHappyPathCommon";
 
 describe("Provision message extension action Dotnet", () => {
   it(
